chore: split up panasonicPTZ functionality, like Pharos is
